+++ new-relic/replay/newrelic-mask-view.js
@@ -2,3 +2,4 @@ import { requireNativeComponent } from 'react-native';
 
 export const NewRelicMask = requireNativeComponent('NRMaskComponentView');
 export const NewRelicUnmask = requireNativeComponent('NRUnMaskComponentView');
+export const NewRelicBlock = requireNativeComponent('NRBlockComponentView');